requirements-api.txt
Flask==3.0.0
Flask-CORS==4.0.0
pandas>=2.0.0
httpx>=0.25.0
aiohttp>=3.9.0
gunicorn==21.2.0
google-cloud-storage>=2.10.0
openai>=1.0.0
tiktoken>=0.5.0
python-dotenv>=1.0.0
numpy>=1.24.0
h5py>=3.8.0
pydantic>=2.0.0
scipy>=1.10.0
networkx>=3.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
faiss-cpu>=1.7.4
tqdm>=4.65.0
